Cleaned up and renamed CheckTaxon
